240308第二次20

+++ rust_web/ws/webservice/src/models/course.rs
@@ -2,10 +2,12 @@ use actix_web::web;
 use chrono::NaiveDateTime;
 use serde::{Deserialize, Serialize};
 
+// use crate::models::course::Course;
+
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Course {
-    pub teacher_id: usize,
-    pub id: Option<usize>,
+    pub teacher_id: i32,
+    pub id: Option<i32>,
     pub name: String,
     pub time: Option<NaiveDateTime>,
 }
